Started cleaning up flask/GCP baseline
