+++ src/flask-diagflow-bot/Dockerfile
@@ -6,6 +6,7 @@ WORKDIR /app
 
 # Copy the current directory contents into the container at /app
 COPY . /app
+COPY ../../data/interim/faq-text-separated.csv /app
 
 # Install any needed packages specified in requirements.txt
 RUN pip install --trusted-host pypi.python.org -r requirements.txt
@@ -13,10 +14,7 @@ RUN pip install python-dotenv
 RUN python -m nltk.downloader wordnet
 
 # Make port 80 available to the world outside this container
-EXPOSE 80
-
-# Define environment variable
-ENV NAME World
+EXPOSE 8080
 
 # Run app.py when the container launches
-CMD ["python", "index.py"]
+CMD ["python", "main.py"]
