--- src/inquire-boulder-bot/requirements.txt
@@ -1,6 +1,0 @@
-Flask
-numpy
-pandas
-dialogflow
-sklearn
-nltk
